+++ rbx_binary/src/deserializer/state.rs
@@ -1,8 +1,7 @@
 use std::{
     collections::{HashMap, HashSet, VecDeque},
     convert::TryInto,
-    io::{self, Read},
-    str,
+    io::Read,
 };
 
 use rbx_dom_weak::{
@@ -15,126 +14,20 @@ use rbx_dom_weak::{
     InstanceBuilder, WeakDom,
 };
 use rbx_reflection::DataType;
-use thiserror::Error;
 
 use crate::{
     cframe,
     chunk::Chunk,
-    core::{
-        find_canonical_property_descriptor, RbxReadExt, FILE_MAGIC_HEADER, FILE_SIGNATURE,
-        FILE_VERSION,
-    },
-    types::{InvalidTypeError, Type},
+    core::{find_property_descriptors, RbxReadExt},
+    types::Type,
 };
 
-/// Represents an error that occurred during deserialization.
-#[derive(Debug, Error)]
-#[error(transparent)]
-pub struct Error {
-    source: Box<InnerError>,
-}
+use super::{error::InnerError, header::FileHeader, Deserializer};
 
-impl From<InnerError> for Error {
-    fn from(inner: InnerError) -> Self {
-        Self {
-            source: Box::new(inner),
-        }
-    }
-}
+pub(super) struct DeserializerState<'a, R> {
+    /// The user-provided configuration that we should use.
+    deserializer: &'a Deserializer<'a>,
 
-#[derive(Debug, Error)]
-pub(crate) enum InnerError {
-    #[error(transparent)]
-    Io {
-        #[from]
-        source: io::Error,
-    },
-
-    #[error("Invalid file header")]
-    BadHeader,
-
-    #[error("Unknown file version {version}. Known versions are: 0")]
-    UnknownFileVersion { version: u16 },
-
-    #[error("Unknown version {version} for chunk {chunk_name}")]
-    UnknownChunkVersion {
-        chunk_name: &'static str,
-        version: u32,
-    },
-
-    #[error(transparent)]
-    InvalidTypeError {
-        #[from]
-        source: InvalidTypeError,
-    },
-
-    #[error(
-        "Type mismatch: Property {type_name}.{prop_name} should be {valid_type_names}, but it was {actual_type_name}",
-    )]
-    PropTypeMismatch {
-        type_name: String,
-        prop_name: String,
-        valid_type_names: &'static str,
-        actual_type_name: String,
-    },
-
-    #[error("Invalid property data: Property {type_name}.{prop_name} was expected to be {valid_value}, but it was {actual_value}")]
-    InvalidPropData {
-        type_name: String,
-        prop_name: String,
-        valid_value: &'static str,
-        actual_value: String,
-    },
-
-    #[error("File referred to type ID {type_id}, which was not declared")]
-    InvalidTypeId { type_id: u32 },
-
-    #[error("Invalid property data: CFrame property {type_name}.{prop_name} had an invalid rotation ID {id:02x}")]
-    BadRotationId {
-        type_name: String,
-        prop_name: String,
-        id: u8,
-    },
-
-    #[error("Expected type id for {expected_type_name} ({expected_type_id:02x}) when reading OptionalCFrame; got {actual_type_id:02x}")]
-    BadOptionalCFrameFormat {
-        expected_type_name: String,
-        expected_type_id: u8,
-        actual_type_id: u8,
-    },
-}
-
-pub(crate) fn decode<R: Read>(reader: R) -> Result<WeakDom, Error> {
-    Ok(decode_inner(reader)?)
-}
-
-pub(crate) fn decode_inner<R: Read>(reader: R) -> Result<WeakDom, InnerError> {
-    let mut deserializer = BinaryDeserializer::new(reader)?;
-
-    loop {
-        let chunk = Chunk::decode(&mut deserializer.input)?;
-
-        match &chunk.name {
-            b"META" => deserializer.decode_meta_chunk(&chunk.data)?,
-            b"SSTR" => deserializer.decode_sstr_chunk(&chunk.data)?,
-            b"INST" => deserializer.decode_inst_chunk(&chunk.data)?,
-            b"PROP" => deserializer.decode_prop_chunk(&chunk.data)?,
-            b"PRNT" => deserializer.decode_prnt_chunk(&chunk.data)?,
-            b"END\0" => {
-                deserializer.decode_end_chunk(&chunk.data)?;
-                break;
-            }
-            _ => match str::from_utf8(&chunk.name) {
-                Ok(name) => log::info!("Unknown binary chunk name {}", name),
-                Err(_) => log::info!("Unknown binary chunk name {:?}", chunk.name),
-            },
-        }
-    }
-
-    Ok(deserializer.finish())
-}
-
-struct BinaryDeserializer<R> {
     /// The input data encoded as a binary model.
     input: R,
 
@@ -166,18 +59,6 @@ struct BinaryDeserializer<R> {
     unknown_type_ids: HashSet<u8>,
 }
 
-/// All the information contained in the header before any chunks are read from
-/// the file.
-pub(crate) struct FileHeader {
-    /// The number of instance types (represented for us as `TypeInfo`) that are
-    /// in this file. Generally useful to pre-size some containers before
-    /// reading the file.
-    pub(crate) num_types: u32,
-
-    /// The total number of instances described by this file.
-    pub(crate) num_instances: u32,
-}
-
 /// Represents a unique instance class. Binary models define all their instance
 /// types up front and give them a short u32 identifier.
 struct TypeInfo {
@@ -203,8 +84,11 @@ struct Instance {
     children: Vec<i32>,
 }
 
-impl<R: Read> BinaryDeserializer<R> {
-    fn new(mut input: R) -> Result<Self, InnerError> {
+impl<'a, R: Read> DeserializerState<'a, R> {
+    pub(super) fn new(
+        deserializer: &'a Deserializer<'a>,
+        mut input: R,
+    ) -> Result<Self, InnerError> {
         let tree = WeakDom::new(InstanceBuilder::new("DataModel"));
 
         let header = FileHeader::decode(&mut input)?;
@@ -212,7 +96,8 @@ impl<R: Read> BinaryDeserializer<R> {
         let type_infos = HashMap::with_capacity(header.num_types as usize);
         let instances_by_ref = HashMap::with_capacity(1 + header.num_instances as usize);
 
-        Ok(BinaryDeserializer {
+        Ok(DeserializerState {
+            deserializer,
             input,
             tree,
             metadata: HashMap::new(),
@@ -224,7 +109,11 @@ impl<R: Read> BinaryDeserializer<R> {
         })
     }
 
-    fn decode_meta_chunk(&mut self, mut chunk: &[u8]) -> Result<(), InnerError> {
+    pub(super) fn next_chunk(&mut self) -> Result<Chunk, InnerError> {
+        Ok(Chunk::decode(&mut self.input)?)
+    }
+
+    pub(super) fn decode_meta_chunk(&mut self, mut chunk: &[u8]) -> Result<(), InnerError> {
         let len = chunk.read_le_u32()?;
         self.metadata.reserve(len as usize);
 
@@ -238,7 +127,7 @@ impl<R: Read> BinaryDeserializer<R> {
         Ok(())
     }
 
-    fn decode_sstr_chunk(&mut self, mut chunk: &[u8]) -> Result<(), InnerError> {
+    pub(super) fn decode_sstr_chunk(&mut self, mut chunk: &[u8]) -> Result<(), InnerError> {
         let version = chunk.read_le_u32()?;
 
         if version != 0 {
@@ -259,7 +148,7 @@ impl<R: Read> BinaryDeserializer<R> {
         Ok(())
     }
 
-    fn decode_inst_chunk(&mut self, mut chunk: &[u8]) -> Result<(), InnerError> {
+    pub(super) fn decode_inst_chunk(&mut self, mut chunk: &[u8]) -> Result<(), InnerError> {
         let type_id = chunk.read_le_u32()?;
         let type_name = chunk.read_string()?;
         let object_format = chunk.read_u8()?;
@@ -300,7 +189,7 @@ impl<R: Read> BinaryDeserializer<R> {
         Ok(())
     }
 
-    fn decode_prop_chunk(&mut self, mut chunk: &[u8]) -> Result<(), InnerError> {
+    pub(super) fn decode_prop_chunk(&mut self, mut chunk: &[u8]) -> Result<(), InnerError> {
         let type_id = chunk.read_le_u32()?;
         let prop_name = chunk.read_string()?;
 
@@ -365,10 +254,14 @@ impl<R: Read> BinaryDeserializer<R> {
         let canonical_name;
         let canonical_type;
 
-        match find_canonical_property_descriptor(&type_info.type_name, &prop_name) {
-            Some(descriptor) => {
-                canonical_name = descriptor.name.clone().into_owned();
-                canonical_type = match &descriptor.data_type {
+        match find_property_descriptors(
+            self.deserializer.database.unwrap(),
+            &type_info.type_name,
+            &prop_name,
+        ) {
+            Some(descriptors) => {
+                canonical_name = descriptors.canonical.name.clone().into_owned();
+                canonical_type = match &descriptors.canonical.data_type {
                     DataType::Value(ty) => *ty,
                     DataType::Enum(_) => VariantType::Enum,
                     _ => {
@@ -1228,7 +1121,7 @@ impl<R: Read> BinaryDeserializer<R> {
         Ok(())
     }
 
-    fn decode_prnt_chunk(&mut self, mut chunk: &[u8]) -> Result<(), InnerError> {
+    pub(super) fn decode_prnt_chunk(&mut self, mut chunk: &[u8]) -> Result<(), InnerError> {
         let version = chunk.read_u8()?;
 
         if version != 0 {
@@ -1260,7 +1153,7 @@ impl<R: Read> BinaryDeserializer<R> {
         Ok(())
     }
 
-    fn decode_end_chunk(&mut self, _chunk: &[u8]) -> Result<(), InnerError> {
+    pub(super) fn decode_end_chunk(&mut self, _chunk: &[u8]) -> Result<(), InnerError> {
         log::trace!("END chunk");
 
         // We don't do any validation on the END chunk. There's no useful
@@ -1272,7 +1165,7 @@ impl<R: Read> BinaryDeserializer<R> {
 
     /// Combines together all the decoded information to build and emplace
     /// instances in our tree.
-    fn finish(mut self) -> WeakDom {
+    pub(super) fn finish(mut self) -> WeakDom {
         log::trace!("Constructing tree from deserialized data");
 
         // Track all the instances we need to construct. Order of construction
@@ -1298,44 +1191,5 @@ impl<R: Read> BinaryDeserializer<R> {
         }
 
         self.tree
-    }
-}
-
-impl FileHeader {
-    pub(crate) fn decode<R: Read>(mut source: R) -> Result<Self, InnerError> {
-        let mut magic_header = [0; 8];
-        source.read_exact(&mut magic_header)?;
-
-        if magic_header != FILE_MAGIC_HEADER {
-            return Err(InnerError::BadHeader);
-        }
-
-        let mut signature = [0; 6];
-        source.read_exact(&mut signature)?;
-
-        if signature != FILE_SIGNATURE {
-            return Err(InnerError::BadHeader);
-        }
-
-        let version = source.read_le_u16()?;
-
-        if version != FILE_VERSION {
-            return Err(InnerError::UnknownFileVersion { version });
-        }
-
-        let num_types = source.read_le_u32()?;
-        let num_instances = source.read_le_u32()?;
-
-        let mut reserved = [0; 8];
-        source.read_exact(&mut reserved)?;
-
-        if reserved != [0; 8] {
-            return Err(InnerError::BadHeader);
-        }
-
-        Ok(Self {
-            num_types,
-            num_instances,
-        })
     }
 }