+++ rbx_binary/src/tests/util.rs
@@ -2,7 +2,7 @@ use std::{fs, path::Path};
 
 use rbx_dom_weak::DomViewer;
 
-use crate::{deserializer::decode, encode, text_deserializer::DecodedModel};
+use crate::{from_reader, text_deserializer::DecodedModel, to_writer};
 
 /// Run a basic gauntlet of tests to verify that the serializer and deserializer
 /// can handle this model correctly.
@@ -28,7 +28,7 @@ pub fn run_model_base_suite(model_path: impl AsRef<Path>) {
 
     // Decode the test file and snapshot a stable version of the resulting tree.
     // This should properly test the deserializer.
-    let decoded = decode(contents.as_slice()).unwrap();
+    let decoded = from_reader(contents.as_slice()).unwrap();
     let decoded_viewed = DomViewer::new().view_children(&decoded);
     insta::assert_yaml_snapshot!(format!("{}__decoded", model_stem), decoded_viewed);
 
@@ -36,7 +36,7 @@ pub fn run_model_base_suite(model_path: impl AsRef<Path>) {
     let decoded_root = decoded.root();
     let top_level_ids = decoded_root.children();
     let mut encoded = Vec::new();
-    encode(&decoded, top_level_ids, &mut encoded).unwrap();
+    to_writer(&mut encoded, &decoded, top_level_ids).unwrap();
 
     // ...but we can snapshot the text representation of what we encoded! In an
     // ideal world, this would be very similar or the same as the text
@@ -51,5 +51,5 @@ pub fn run_model_base_suite(model_path: impl AsRef<Path>) {
     // We don't make any assertions about the result right now, as our format
     // support is still lacking. In the future, we should assert that this is
     // the same as the original decoding of the test file.
-    decode(encoded.as_slice()).unwrap();
+    from_reader(encoded.as_slice()).unwrap();
 }