+++ rbx_types/src/variant.rs
@@ -25,7 +25,6 @@ macro_rules! make_variant {
         #[cfg_attr(
             feature = "serde",
             derive(serde::Serialize, serde::Deserialize),
-            serde(tag = "Type", content = "Value"),
         )]
         pub enum Variant {
             $(
@@ -144,21 +143,17 @@ mod serde_test {
         let vec2 = Variant::Vector2(Vector2::new(5.0, 7.0));
 
         let ser = serde_json::to_string(&vec2).unwrap();
-        assert_eq!(ser, r#"{"Type":"Vector2","Value":[5.0,7.0]}"#);
+        assert_eq!(ser, r#"{"Vector2":[5.0,7.0]}"#);
 
         let de: Variant = serde_json::from_str(&ser).unwrap();
         assert_eq!(de, vec2);
     }
 
     #[test]
-    #[ignore]
     fn non_human() {
         let vec2 = Variant::Vector2(Vector2::new(5.0, 7.0));
 
         let ser = bincode::serialize(&vec2).unwrap();
-
-        // FIXME: This call currently fails because bincode does not support
-        // Deserializer::deserialize_identifier.
 
         let de: Variant = bincode::deserialize(&ser).unwrap();
         assert_eq!(de, vec2);